fixed a couple of compilation warnings

+++ ColorUtilities.cpp
@@ -59,10 +59,10 @@ LookupTable ColorUtilities::init_glasbey() {
             + GLASBEY_FILE);
 
     while (!f.eof()) {
-        int l = 0;
-        int r = 0;
-        int g = 0;
-        int b = 0;
+        uint8_t l = 0;
+        uint8_t r = 0;
+        uint8_t g = 0;
+        uint8_t b = 0;
 
         f >> l >> r >> g >> b; // reads a line from the file, interpeting 
                                // the first number as label and the following
